backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.0.3
transformers==4.44.2
torch>=2.0.0,<3.0.0
gunicorn==21.2.0

